+++ spec/services/ai_concierge_v3/orchestration/core/response_payload_builder_spec.rb
@@ -1,5 +1,5 @@
 require "rails_helper"
-RSpec.describe AiConciergeV3::Orchestration::Result do
+RSpec.describe AiConciergeV3::Orchestration::Core::ResponsePayloadBuilder do
   it "exists" do
     expect(described_class).to be_a(Class)
   end