+++ spec/services/ai_concierge_v3/orchestration/core/result_spec.rb
@@ -1,5 +1,5 @@
 require "rails_helper"
-RSpec.describe AiConciergeV3::Orchestration::ResponsePayloadBuilder do
+RSpec.describe AiConciergeV3::Orchestration::Core::Result do
   it "exists" do
     expect(described_class).to be_a(Class)
   end